Backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Darshan <karkeradarshan98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.8.1"
azure-ai-textanalytics = "5.3.0"
azure-cognitiveservices-speech = "1.41.1"
azure-common = "1.1.28"
azure-core = "1.31.0"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
django = "5.1.2"
idna = "3.10"
urllib3 = "2.2.3"
tzdata = "2024.2"
typing-extensions = "4.12.2"
python-dotenv = "1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
